STILL Debugging
Fixed problems (hopefully for good) with f64/sample converters. Working on a problem with samples being inflated with 0s.

+++ src/project/wave.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Clone, PartialEq, Copy, Default)]
 pub struct Wave {
-    pub freq: f32,
-    pub amp: f32,
-    pub phase_shift: f32
+    pub freq: f64,
+    pub amp: f64,
+    pub phase_shift: f64
 }
